test(admin): cover pool CLI dispatch
